Combine sum horizontal tests
