+++ tests/expr/sum_horizontal_test.py
@@ -16,3 +16,17 @@ def test_sumh(constructor: Any) -> None:
         "horizonal_sum": [5, 7, 8],
     }
     compare_dicts(result_native, expected)
+
+
+def test_sumh_literal(constructor: Any) -> None:
+    data = {"a": [1, 3, 2], "b": [4, 4, 6], "z": [7.0, 8, 9]}
+    df = nw.from_native(constructor(data), eager_only=True)
+    result = df.with_columns(horizonal_sum=nw.sum_horizontal("a", nw.col("b")))
+    result_native = nw.to_native(result)
+    expected = {
+        "a": [1, 3, 2],
+        "b": [4, 4, 6],
+        "z": [7.0, 8.0, 9.0],
+        "horizonal_sum": [5, 7, 8],
+    }
+    compare_dicts(result_native, expected)
